Move src to services/src to improve build context
There are dependencies between services and src, which required a
context that included both directories. Now that src is under services,
we can make services the context.

There are some large directories created during builds that don't
need to be part of the context, so a .dockerignore file has been
created.