Remove unused dependency on javafx.util.pair, extend pipline
The javafx.util.pair wasn't used, and it was causing a compilation
failure in maven:3-jdk-8 container. Not sure why, as it is suppose to be
part of java8 by standard. But, pair wasn't used anyway, so removed it.

With that, we can add mvn -f services/src and services/wfm.