Leverage External ISL DB to filter checks
The goal is to define a filter that is used to determine, once a port up
event is received, whether we should send a discovery packet through it
or not.

The filter(real one) implementation will most likely reach out to an
external system (maybe through a rest API) to determine whether an ISL
discovery packet should be sent. Now we add interface for such filter
and it's dummy implementation to check how it works.

Closes #61