Closes Issue #79 - kafka topics refactor (#120)
* DOC: updated kafka design doc

* MAINT: Refactor kafka topics in Floodlight Module

This is part of getting rid of kilda-test.

Initial checkin refactors kilda-test out of the floodlight module.

* MAINT: topology_engine, fl-modules kafka changes

Still in the process of Kafka Topic Changes

* MAINT: Refactor kafka topics. Intermediary update

This commit holds a lot of fairly minor changes to enable
the refactoring of the kafka topics.

NB: There are still some unresolved unit tests, primarily in the
flow tests.

* MAINT: Unit Tests for Kafka Topics change passing

This checkin completes the change to revamp the kafka topics.
NB: Whereas the unit tests pass, the acceptance tests will most likely
fail. That'll be the subject of future issues / checkins.