Add BaseMessage class to all messages
This unifies all messages to have a common base class.
The current base class has a timestamp, which may not be the best
name and should be changed to message_timestamp, or something like
that.

As part of this change, I've removed the Subtype tracking, and instead
rely on the configuration of BaseMessage which includes the class
name in the serialized json, which facilitates the deserialization.