MAINT: Additional Message Refactoring (#135)
* DOC: updated kafka design doc

* MAINT: Refactor kafka topics in Floodlight Module

This is part of getting rid of kilda-test.

Initial checkin refactors kilda-test out of the floodlight module.

* MAINT: topology_engine, fl-modules kafka changes

Still in the process of Kafka Topic Changes

* MAINT: Refactor kafka topics. Intermediary update

This commit holds a lot of fairly minor changes to enable
the refactoring of the kafka topics.

NB: There are still some unresolved unit tests, primarily in the
flow tests.

* MAINT: Unit Tests for Kafka Topics change passing

This checkin completes the change to revamp the kafka topics.
NB: Whereas the unit tests pass, the acceptance tests will most likely
fail. That'll be the subject of future issues / checkins.

* Add BaseMessage class to all messages

This unifies all messages to have a common base class.
The current base class has a timestamp, which may not be the best
name and should be changed to message_timestamp, or something like
that.

As part of this change, I've removed the Subtype tracking, and instead
rely on the configuration of BaseMessage which includes the class
name in the serialized json, which facilitates the deserialization.

* MAINT: Add BaseMessage and fix Speaker test

BaseMessage.java didn't make it into the last commit.

And, the tests in Speaker weren't updated to reflect changed
message format.

There is still code in python that needs updating.