[MAINT] Add ping to mininet_rest with test
mininet_rest now has a /pingable endpoint, which relies
on a strategy used to test flows - ie that we have
a linear set of switches, numbered 00000001 to 00000005.

Work could be done to move this out of mininet_rest and into
some other place, ie the Java side, but that is an exercise for
another day. We'd have to re-think what we can rely on, and what
would have to be discovered.