[MAINT] adjust flow test topology - uniformity
The previous topology had 5 nodes, the new one has 7.
The reason for the change is to increase uniformity in test code/hacks.

The code already had special logic if the switch was #1 (only one link)
But the code *needed* special logic for switch #5, which also has
only one link. This special logic was missing.  Rather than add it,
I changed the topology so that special logic wasn't needed.