Add flow segments for flows; FIX avail..bandwidth
ISL available.bandwidth was slightly broken in that it calculated it
based on the forward path .. which inadvertently does the right thing
today, but not tomorrow.  Because each node is an endpoint of an ISL,
and the old algorithm deducted BW for each node (endpoint), it
correctly adjusted the BW for an ISL to 2x - since initial algorithm
as reverse = opposite of forward.

However, if we change this algorithm and allow for asymmetric flows,
then the old algorithm would do the wrong thing.

Additionally, added :flow_segment relationships so that we can easily
track the entire path of a flow, the cookie for each segment if we
decide to change that. Additionally, we can accurately account for the
hit to the available_bw based on a dynamic query, rather than
manually adding / subtracting the available_bw.