Fix logic flaw in deleting flows in TE
The old logic used segment flows if not a single switch flow.
However, this left us with a dangling rule at the end of the flow.
New logic is simplified - all flows (single / multi) need to
delete the rule on the src_switch. Then, if there are any segments,
delete the rule on the dst_switch per segment.