Update kafka's spout configuration
First issue .setMaxUncommittedOffsets(Integer.MAX_VALUE) set extremelly big
backlog size. This can lead to uncontrollable memory consumption. Deagult is
big enought for any our task(10_000_000).

This second option replaced with default value is

.setRetry(new KafkaSpoutRetryExponentialBackoff(
	TimeInterval.seconds(5), TimeInterval.microSeconds(5),
	Integer.MAX_VALUE, TimeInterval.seconds(60)))

It is almost match with default value. The only difference in initial timeout -
in default value it is lower. So it is safe to drop replacement of this option
too.