Implement tools that allow to catch errors returned via OF
Add new service named SessionService. It is responsible for "creation" and
tracking communication "session" with connected switches. The session itself is
a closable object that can/should be used into try-with-resources construction.
All messages send via session will be traced for possible error responses.
Session will send barrier request when it closed.

So when we receive barrier response we can treat all sent messages as
successfully processed (if we do not receive error response for them before).

Using this session we can catch and handle error responses have been sent by
switch on our OF requests.