Change ThrotthlingBolt parallelism to 1
Reroute throtthling should be capable for suspending
reroute of all affected flows in case of network flapping.
However due to parallelism of ThrotthlingBolt=2 and
field grouping by flow id it's not the case. Since only
50% of flows are threated by one worker.
This patch change this level of parallelism as a hotfix.
While it should be refactored so it will be possible to
set level of parallelism more than 1.

RelatedTo: #1636