Change MIN_METER_ID to 32.
At this moment MIN_METER_ID for flow meters is 11.
This value is used to allocate meter IDs for flows.
But also we need to allocate meter IDs for default rules.
To do it special mask 'PACKET_IN_RULES_METERS_MASK' is used.
With the help of this mask we take last 5 bits of
default rule cookie to create meter ID.
That means we have range 1..31 for default rule meter IDs.
MIN_METER_ID is used to do not intersect flow meter IDs with this range.

Closes #1801