Use real SwitchManagerConfig in unit tests
We have implementation SwitchManagerConfig for tests with hardcoded values.
The problem is that this config is used as expected config and as actual config.
Some time ago flowMeterBurstCoefficient was changed from 0.1 to 1.05
but tests still use old 0.1 value. To fix it missed @Default annotations were added
to the real config class and now this class is used in tests.
Fake config class for tests was moved to separate class to store expected values
and fail if actual values will be changed.