Lower log level for "lost" discovery packets
Because discovery packets are send via all "enabled" ports and because
not all "enabled" port have an ISL on them we are loosing a lot of
discovery packets.

On production scale system log message about lost discovery packet will
be produced thouthant times per minute. It is extremelly noisy. So
lowering level for this message to debug.

Plus fix couple of typos.