Fixed shouldSendDatapointRequestsTwice test.
This test expects to see 2 POST requests into OpenTsdb because we want
to store 2 metrics. But if these 2 metrics will be in one batch only 1
http request will be send (with 2 metrics inside). This test worked previosly
because parallelist of OtsdbBolt was 10 and we have 10 different batches.
Now parallelist is only 1 so Bolt puts both metrics into one batch and send one
request. To fix it batch size was reduced to 1 for this test.