fix max latency reason not shown
In the case the path is not fully calculated, we can't know the latency of the
full path because there is a point during the path computation where the latency
is already higher than the max_latency property.

For this case the reason was not properly added so it wasn't inlcuded in
the response.

This commit fix it.