fix findOverLargeIslands unit test
During the #2910 pull request, this test should have failed. However,
it was overlooked because it was being ignored. This PR change the way
the paths were stored, so the path segments became half the size.

This commit updates the segment size assertion in half and unignore
the test since it is not taking long.

Closes #5118