feat(ha-flow-ping): add ha flow periodic pings
This commit introduces the HA Flow periodic ping feature, enabling
the ability to periodically check HA flows that are flagged with
'periodic_pings' set to true. This functionality operates similarly
to the current periodic ping for Flows and Y-Flows.

The ping results are stored in a time-series database, with the
default metric being "kilda.flow.latency" and tagged with the
corresponding "ha_flow_id".

Part of #5153