bugfix: add rule to send ping packet to transit This commit updates the functionality of skipping a ping packet in the egress table. Now the packet that doesn't belong to this switch is sent directly to the transit table.
This is done in the ingress table. To achieve this, the current
EgressIslVlanRuleGenerator functionality has been copied, but with the
addition of matching the eth_src=ping_magic_mac and sending the packet to transit.

This change was necessary because the current hardware switches do not support
matching src_mac and dst_mac in tables other than 0.

Closes #5232