fix switch synchronization when discrepancy doesn't include cookie
Switch sync fails when there are misconfigured rules with metadata and the cookie
is the same for the expected rule and the misconfigured rule. Since the list of
misconfigured rules contains a cookie for the misconfigured rule, and the
misconfigured rules has the correct cookie, the cookie in this list is null
causing a null pointer when the validator tries to get it.