Add a delta to compare all the metric rates
There is a physical switch that stores the metric rate number using a floating
point variable. This means that sometimes the rate value is not properly compared
with the expected value.

This commit applies a max delta to all the metric rates comparations that allow
to introduce an expected error during the comparison.

Closes #5638